+++ libft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: fbougama <fbougama@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/08 15:33:03 by fbougama          #+#    #+#             */
-/*   Updated: 2019/10/10 13:24:36 by fbougama         ###   ########.fr       */
+/*   Updated: 2019/10/15 13:44:51 by fbougama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ char			*ft_strnstr(const char *haystack, const char *needle, size_t n)
 	size_t	needle_len;
 	int		i;
 	size_t	haystack_len;
-	int		j;
+	size_t	j;
 
 	needle_len = ft_strlen((char*)needle);
 	haystack_len = ft_strlen((char*)haystack);
